Split-form-nd-updating-friend-list-with-new-balance-added
