--- src/App.js
@@ -1,172 +1,0 @@
-import { Children, useState } from "react";
-
-const initialFriends = [
-  {
-    id: 118836,
-    name: "Clark",
-    image: "https://i.pravatar.cc/48?u=118836",
-    balance: -7,
-  },
-  {
-    id: 933372,
-    name: "Sarah",
-    image: "https://i.pravatar.cc/48?u=933372",
-    balance: 20,
-  },
-  {
-    id: 499476,
-    name: "Anthony",
-    image: "https://i.pravatar.cc/48?u=499476",
-    balance: 0,
-  },
-];
-
-export default function App() {
-  const [friends, setFriends] = useState(initialFriends);
-  const [showAddFriend, setShowAddFriend] = useState(false);
-  const [selectedFriend, setSelectedFriend] = useState(null);
-
-  function handleAddFriend(newFriend) {
-    setFriends([...friends, newFriend]);
-    setShowAddFriend(!showAddFriend);
-  }
-
-  function handleShowAddFriend() {
-    setShowAddFriend(!showAddFriend);
-  }
-  function handleSelection(friend) {
-    setSelectedFriend(() => (friend === selectedFriend ? null : friend));
-    setShowAddFriend(false);
-  }
-  return (
-    <div className="app">
-      <div className="sidebar">
-        <FriendsList
-          friends={friends}
-          selectedFriend={selectedFriend}
-          onSelectedFriend={handleSelection}
-        />
-
-        {showAddFriend && <FormAddFriend onAddFriends={handleAddFriend} />}
-
-        <Button onClick={handleShowAddFriend}>
-          {showAddFriend ? "Close" : "Add friend"}
-        </Button>
-      </div>
-
-      {selectedFriend && <FormSplitBill selectedFriend={selectedFriend} />}
-    </div>
-  );
-}
-
-function FriendsList({ friends, selectedFriend, onSelectedFriend }) {
-  return (
-    <ul>
-      {friends.map((friend) => (
-        <Friend
-          friend={friend}
-          key={friend.id}
-          selectedFriend={selectedFriend}
-          onSelectedFriend={onSelectedFriend}
-        />
-      ))}
-    </ul>
-  );
-}
-
-function Friend({ friend, selectedFriend, onSelectedFriend }) {
-  const isSelected = selectedFriend === friend;
-  return (
-    <li className={isSelected ? "selected" : ""}>
-      <img src={friend.image} alt={friend.name} />
-      <h3>{friend.name}</h3>
-      {friend.balance < 0 && (
-        <p className="red">
-          You owe {friend.name} {Math.abs(friend.balance)}€
-        </p>
-      )}
-      {friend.balance > 0 && (
-        <p className="green">
-          {friend.name} owes you {Math.abs(friend.balance)}€
-        </p>
-      )}
-      {friend.balance === 0 && <p>You and {friend.name} are even</p>}
-      <Button onClick={() => onSelectedFriend(friend)}>
-        {isSelected ? "Close" : "Select"}
-      </Button>
-    </li>
-  );
-}
-
-function Button({ children, onClick }) {
-  return (
-    <button className="button" onClick={onClick}>
-      {children}
-    </button>
-  );
-}
-
-function FormAddFriend({ onAddFriends }) {
-  const [name, setName] = useState("");
-  const [image, setImage] = useState("https://i.pravatar.cc/48");
-  function handleSubmit(e) {
-    if (!name || !image) return;
-
-    e.preventDefault();
-    const id = crypto.randomUUID();
-    const newFriend = {
-      name,
-      image: `${image}?=${id}`,
-      balance: 0,
-      id,
-    };
-    onAddFriends(newFriend);
-
-    setName("");
-    setImage("https://i.pravatar.cc/48");
-  }
-  return (
-    <form className="form-add-friend" onSubmit={handleSubmit}>
-      <label>👭 Friend name</label>
-      <input
-        type="text"
-        value={name}
-        onChange={(e) => setName(e.target.value)}
-      />
-
-      <label>🌇 Image URL</label>
-      <input
-        type="text"
-        value={image}
-        onChange={(e) => setImage(e.target.value)}
-      />
-
-      <Button>Add</Button>
-    </form>
-  );
-}
-
-function FormSplitBill({ selectedFriend }) {
-  return (
-    <form className="form-split-bill">
-      <h2>Split a bill with {selectedFriend.name}</h2>
-
-      <label>💰 Bill value</label>
-      <input type="text" />
-
-      <label>🧍‍♀️ Your expense</label>
-      <input type="text" />
-
-      <label>👭 {selectedFriend.name}'s expense</label>
-      <input type="text" disabled />
-
-      <label>🧍‍♀️ Who is paying the bill</label>
-      <select>
-        <option value="user">You</option>
-        <option value="friend">{selectedFriend.name}</option>
-      </select>
-
-      <Button>Add</Button>
-    </form>
-  );
-}
