+++ src/components/App.js
@@ -1,0 +1,79 @@
+import { useState } from "react";
+import { FormSplitBill } from "./FormSplitBill";
+import { FormAddFriend } from "./FormAddFriend";
+import { Button } from "./Button";
+import { FriendsList } from "./FriendsList";
+
+const initialFriends = [
+  {
+    id: 118836,
+    name: "Clark",
+    image: "https://i.pravatar.cc/48?u=118836",
+    balance: -7,
+  },
+  {
+    id: 933372,
+    name: "Sarah",
+    image: "https://i.pravatar.cc/48?u=933372",
+    balance: 20,
+  },
+  {
+    id: 499476,
+    name: "Anthony",
+    image: "https://i.pravatar.cc/48?u=499476",
+    balance: 0,
+  },
+];
+
+export default function App() {
+  const [friends, setFriends] = useState(initialFriends);
+  const [showAddFriend, setShowAddFriend] = useState(false);
+  const [selectedFriend, setSelectedFriend] = useState(null);
+
+  function handleAddFriend(newFriend) {
+    setFriends([...friends, newFriend]);
+    setShowAddFriend(!showAddFriend);
+  }
+
+  function handleShowAddFriend() {
+    setShowAddFriend(!showAddFriend);
+  }
+  function handleSelection(friend) {
+    setSelectedFriend(() => (friend === selectedFriend ? null : friend));
+    setShowAddFriend(false);
+  }
+  function handleSplitBill(value) {
+    setFriends((friends) =>
+      friends.map((friend) =>
+        friend.id === selectedFriend.id
+          ? { ...friend, balance: friend.balance + value }
+          : friend
+      )
+    );
+    setSelectedFriend(null);
+  }
+  return (
+    <div className="app">
+      <div className="sidebar">
+        <FriendsList
+          friends={friends}
+          selectedFriend={selectedFriend}
+          onSelectedFriend={handleSelection}
+        />
+
+        {showAddFriend && <FormAddFriend onAddFriends={handleAddFriend} />}
+
+        <Button onClick={handleShowAddFriend}>
+          {showAddFriend ? "Close" : "Add friend"}
+        </Button>
+      </div>
+
+      {selectedFriend && (
+        <FormSplitBill
+          selectedFriend={selectedFriend}
+          onSplitBill={handleSplitBill}
+        />
+      )}
+    </div>
+  );
+}
