+++ src/components/FriendsList.js
@@ -1,0 +1,16 @@
+import { Friend } from "./Button";
+
+export function FriendsList({ friends, selectedFriend, onSelectedFriend }) {
+  return (
+    <ul>
+      {friends.map((friend) => (
+        <Friend
+          friend={friend}
+          key={friend.id}
+          selectedFriend={selectedFriend}
+          onSelectedFriend={onSelectedFriend}
+        />
+      ))}
+    </ul>
+  );
+}
